pyproject.toml
[tool.poetry]
name = "ddpy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.10.1"
pylint = "^2.6.0"
boto3 = "^1.16.3"
requests = "^2.24.0"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
run = "ddpy:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
